# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gate/check:
    post:
      operationId: checkGate
      tags: [progress]
      summary: Check if content is accessible (gate validation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateCheckInput"
      responses:
        "200":
          description: Gate status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateCheckResult"

  # Progress
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List available exams with unlock status
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [topic, chapter, subject, dpp, pyq, grand]
        - name: subjectId
          in: query
          schema:
            type: string
        - name: chapterId
          in: query
          schema:
            type: string
        - name: topicId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create an exam (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam detail
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exam detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
  /exams/{examId}/start:
    post:
      operationId: startExam
      tags: [exams]
      summary: Start an exam attempt
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exam started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamAttempt"
        "403":
          description: Gate prerequisite not met
  /attempts/{attemptId}/answer:
    post:
      operationId: saveAnswer
      tags: [exams]
      summary: Save answer for a question
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /attempts/{attemptId}/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam attempt
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExamInput"
      responses:
        "200":
          description: Exam submitted with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /attempts/{attemptId}/pause:
    post:
      operationId: pauseExam
      tags: [exams]
      summary: Pause exam attempt
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /attempts/{attemptId}/resume:
    post:
      operationId: resumeExam
      tags: [exams]
      summary: Resume exam attempt
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamAttempt"
  /results/{resultId}:
    get:
      operationId: getResult
      tags: [exams]
      summary: Get exam result and analysis
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth schemas
...
    GateCheckInput:
      type: object
      properties:
        targetId:
          type: string
        targetType:
          type: string
          enum: [lecture_quiz, dpp, pyq, topic_test, chapter_test, subject_test, pdf_upload]
      required:
        - targetId
        - targetType
    GateCheckResult:
      type: object
      properties:
        allowed:
          type: boolean
        nextStep:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
      required:
        - allowed
        - gateStatus

    # Progress
...
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [lecture_quiz, dpp, pyq, topic_test, chapter_test, subject_test, grand_test]
        subjectId:
          type: ["string", "null"]
        chapterId:
          type: ["string", "null"]
        topicId:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        totalQuestions:
          type: integer
        totalMarks:
          type: integer
        passingScore:
          type: ["integer", "null"]
        negativeMarking:
          type: number
        isUnlocked:
          type: boolean
        hasAttempted:
          type: boolean
        lastScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - durationMinutes
        - totalQuestions
        - totalMarks
        - negativeMarking
        - isUnlocked
        - hasAttempted
    ExamInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [lecture_quiz, dpp, pyq, topic_test, chapter_test, subject_test, grand_test]
        subjectId:
          type: string
        chapterId:
          type: string
        topicId:
          type: string
        durationMinutes:
          type: integer
        passingScore:
          type: integer
        negativeMarking:
          type: number
        questionIds:
          type: array
          items:
            type: string
      required:
        - title
        - type
        - durationMinutes
    ExamDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        durationMinutes:
          type: integer
        totalQuestions:
          type: integer
        totalMarks:
          type: integer
        passingScore:
          type: ["integer", "null"]
        negativeMarking:
          type: number
        isUnlocked:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - id
        - title
        - type
        - durationMinutes
        - totalQuestions
        - totalMarks
        - negativeMarking
        - isUnlocked
        - questions
    ExamAttempt:
      type: object
      properties:
        id:
          type: string
        examId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [in_progress, paused, submitted, auto_submitted]
        startTime:
          type: string
          format: date-time
        pauseCount:
          type: integer
        remainingSeconds:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AttemptAnswer"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - id
        - examId
        - userId
        - status
        - startTime
        - pauseCount
        - remainingSeconds
        - answers
        - questions
    AttemptAnswer:
      type: object
      properties:
        questionId:
          type: string
        selectedOption:
          type: ["integer", "null"]
        isMarkedForReview:
          type: boolean
        timeSpentSeconds:
          type: integer
      required:
        - questionId
        - isMarkedForReview
        - timeSpentSeconds
    AnswerInput:
      type: object
      properties:
        questionId:
          type: string
        selectedOption:
          type: ["integer", "null"]
        isMarkedForReview:
          type: boolean
        timeSpentSeconds:
          type: integer
      required:
        - questionId
        - isMarkedForReview
        - timeSpentSeconds
    SubmitExamInput:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required:
        - answers
    ExamResult:
      type: object
      properties:
        id:
          type: string
        examId:
          type: string
        examTitle:
          type: string
        examType:
          type: string
        score:
          type: number
        maxScore:
          type: number
        accuracy:
          type: number
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        incorrectAnswers:
          type: integer
        skippedAnswers:
          type: integer
        timeTakenSeconds:
          type: integer
        percentile:
          type: ["number", "null"]
        passed:
          type: boolean
        questionWise:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
        topicWise:
          type: array
          items:
            $ref: "#/components/schemas/TopicWiseResult"
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - examId
        - examTitle
        - examType
        - score
        - maxScore
        - accuracy
        - totalQuestions
        - correctAnswers
        - incorrectAnswers
        - skippedAnswers
        - timeTakenSeconds
        - passed
        - questionWise
        - topicWise
        - submittedAt
...
    QuestionResult:
      type: object
      properties:
        questionId:
          type: string
        questionText:
          type: string
        selectedOption:
          type: ["integer", "null"]
        correctOption:
          type: integer
        isCorrect:
          type: boolean
        marksAwarded:
          type: number
        timeSpentSeconds:
          type: integer
        textSolution:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        qrCodeSvg:
          type: ["string", "null"]
      required:
        - questionId
        - questionText
        - correctOption
        - isCorrect
        - marksAwarded
        - timeSpentSeconds
    TopicWiseResult:
      type: object
      properties:
        topicId:
          type: string
        topicName:
          type: string
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        accuracy:
          type: number
      required:
        - topicId
        - topicName
        - totalQuestions
        - correctAnswers
        - accuracy

    # Questions
    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: ["integer", "null"]
        marks:
          type: number
        topicId:
          type: ["string", "null"]
        textSolution:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        qrCodeSvg:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required:
        - id
        - text
        - options
        - marks
        - difficulty
...