# lib/api-zod/src/generated/types/gateCheckResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EdTech Study Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GateCheckResultGateStatus } from './gateCheckResultGateStatus';

export interface GateCheckResult {
  allowed: boolean;
  /** @nullable */
  nextStep?: string | null;
  /** @nullable */
  reason?: string | null;
  gateStatus: GateCheckResultGateStatus;
}